Couple Tweaks
Added a point called Helipoint that may be used for the helicopter,
rather than having to use hard coordinates, may need to add a status
like "flying" or something to keep the helicopter from spawning in the
water. Also fixed several arrays that hadn't been updated from the old
sandbox.